fix: wire widths, 32-sim overflows and stale limits (NET_PROTOCOL_VER 88)
WIRE FORMAT CHANGES — proto 87 -> 88. Old peers mis-parse all three, so
server and clients must be deployed together; mixed versions are rejected at
auth with DISCONNECT_VERSION, which is the intended outcome.

SVC_ELIM_STATE — the eliminated set was a u16 over 48 slots
  Simulants wire-key at MAX_PLAYERS+ordinal (16..47), so every sim's
  eliminated flag truncated away: the host marked a sim out and no client ever
  agreed, leaving the lives HUD, elimChrCanRespawn and the match-end condition
  permanently disagreeing. `1u << i` for i >= 32 was also undefined behaviour,
  executed on both sides every heartbeat. Now one byte per combatant, appended
  after the team-pool block; elimApplyWireState takes an array instead of a
  mask.

CLC_PICKUP_REQUEST / CLC_DOOR_ACTIVATE / SVC_CARRY_STATE — syncid truncation
  prop->syncid is u32 and g_NetNextSyncId free-runs for the whole stage, so a
  long high-churn round passes 65535. These three marshalled it as u16, so the
  server resolved a DIFFERENT prop (id 70000 -> 4464) and then ran
  objTestForPickup / propdoorInteract on it as the requesting client. All three
  are now u32. The server-side hit-validation ring (srvhits[].syncid) had the
  same width bug — a shot at prop 70000 validated against prop 4464 — and is
  now u32 too. SVC_PROP_RECONCILE keeps its u16 field: that one already has an
  explicit, documented NET_RECONCILE_MAXSYNCID guard.

netvotecandidate.playlist_index — s8 against a 256-entry playlist
  Entry 130 wrapped to -56 and entry 255 to exactly -1, which is the RANDOM
  sentinel: the ballot slot rendered as "Random" and the vote resolved to a
  freshly-picked entry, so entries past 127 could never be voted for. Now s16
  with an 0xFFFF wire sentinel; playlistPickBallot and its caller widened to
  match.

NON-WIRE

playlist.h — `u8 count` made the overflow guard dead code
  PLAYLIST_MAX_ENTRIES is 256, so `count >= 256` on a u8 is always false: the
  "too many entries" warning never printed, count++ wrapped 255 -> 0, and a
  long playlist silently overwrote its own first entries. Widened to s16.

playlist.c — the scenario table was three scenarios behind
  s_scenarios stopped at CAPTURETHECASE, so `scenario = ZONES` in a playlist
  file fell through parseScenario's strtol fallback to 0 = Combat with no
  warning, and scenarioName returned "?" — which saverotation wrote to disk and
  the next load parsed back as Combat. RANDOM also hardcoded `% 6u` so it could
  never pick them. Added GRAFFITI/PAINTROOM/PAINT/ZONES/RACE to the parse
  table, the three reverse-lookup cases, and switched RANDOM to
  MPSCENARIO_COUNT.

challenge.c — two stale limits
  `chrslots &= 0x000f` is a 4-player literal that runs every tick from
  menutick's MPSETUP block, so a 5+ player lobby lost players 4-15 from the
  Teams menus for the rest of the session (the sibling branch at the bottom of
  the same function was already netplay-corrected). And
  g_MpSimulantDifficultiesPerNumPlayers rows are MAX_LOCAL_PLAYERS (4) wide
  while numplayers is counted over MAX_PLAYERS (16) chrslot bits, so 5+ clients
  indexed past the row and read a neighbouring bot's difficulty — silently
  changing sims' difficulty or disabling them, since BOTDIFF_DISABLED also
  drives the chrslots bit. Clamped to the range the table actually describes.

mplayer.c — bot-name sprintf overflowed onto mpheadnum
  base.name is char[15] with mpheadnum packed at 0x0f. The longest profile
  string is 11 chars ("VendettaSim"), so a two-digit count gives
  "VendettaSim:10\n" = 15 chars + NUL = 16 bytes, and the strcpy put that NUL
  on mpheadnum, zeroing the sim's configured head. Upstream capped at 8 sims
  (single digit, exactly 15 bytes); the 32-simulant bump made it reachable.

elimination.inc — 32-faction cap vs 36 possible factions
  In free-for-all elimFactionIndex returns the combatant's own packed index,
  which reaches MAX_MPCHRS (48) offline with 1 human + 32 sims. The u32 `seen`
  mask skipped factions 32+, so the match ended while combatants with lives
  remaining were still alive. Widened to u64.

race.inc — packed-vs-slot index confusion
  raceTick writes nextcp/lapsdone/finishpos indexed by the PACKED
  g_MpAllChrPtrs position, but raceCalculatePlayerScore is handed a chrslots
  SLOT index, so it read another combatant's progress whenever the two differ.
  Now converts via func0f18d0e8, exactly as htmCalculatePlayerScore does. Same
  class as the packed-vs-slot bug recorded for proto 82.

spectator.c — playerorder written with an unbounded value
  The panel loop bounded the write INDEX but not the value: playerorder entries
  are used as g_Vars.players[playerorder[i]], and on a full dedicated server
  `combatants` reaches MAX_PLAYERS, so combatants+i indexed one past the end.
  (panels is forced to >= 1 before the dedicated early-return, so this path
  runs there.)

bondgun.c — the viewmodel FOV slider moved remote players' muzzles
  bgunGetRenderFovY feeds bgunGetFovOffsetY/Z, which shift the hand
  translation that becomes hand->muzzlepos — the literal spawn position of
  fired projectiles. A remote pawn's bgunTick runs locally, so a purely
  cosmetic local slider (and chaos "WAYTOODANK Viewmodel" at 170 deg, worth
  ~37 units of forward displacement) displaced their projectile origins; worse,
  PLAYER_EXTCFG() indexes by `mpindex & 3`, so a listen host derived a remote
  client's rocket origin from a local splitscreen player's slider. Remote pawns
  now use PLAYER_DEFAULT_FOV.

playermgr.c — removed a shipped debug probe
  A self-labelled `// PROBE (remove after Fix #4)` netDiagLogf was running
  ungated in the stage-load path.

Compile-verified (GCC/MinGW x64, exit 0, zero diagnostics in the touched
files). Runtime PENDING.

+++ port/include/net/playlist.h
@@ -52,7 +52,10 @@ struct playlistentry {
 
 struct playlist {
 	struct playlistentry entries[PLAYLIST_MAX_ENTRIES];
-	u8 count;
+	// s16, not u8: PLAYLIST_MAX_ENTRIES is 256, so `count >= 256` on a u8 is
+	// always false — the "too many entries" guard was unreachable, count++
+	// wrapped 255 -> 0, and a long playlist silently overwrote its own start.
+	s16 count;
 
 	// Server header settings — read from the [server] section of the
 	// playlist file. vote_candidates is clamped to [2, num_entries]; if the
@@ -91,7 +94,7 @@ s32 playlistPick(const struct playlist *pl, u64 *rng_state);
 // per-entry weights but guarantees no duplicates. If pl->random_in_pool is
 // set and n > 1, the last slot is reserved for a sentinel RANDOM choice that
 // the caller resolves at apply time.
-s32 playlistPickBallot(const struct playlist *pl, u64 *rng_state, s32 n, s8 *out_indices);
+s32 playlistPickBallot(const struct playlist *pl, u64 *rng_state, s32 n, s16 *out_indices);
 
 // Resolve RANDOM sentinels in the entry into concrete picks (stage, scenario,
 // preset). The output is a copy of the input with sentinels replaced; the
